+++ packages/react/src/experimental/Modals/OneModal/OneModal.tsx
@@ -1,7 +1,7 @@
 import { TabsProps } from "@/experimental/Navigation/Tabs"
-import { cn } from "@/lib/utils"
-import { Dialog, DialogContent } from "@/ui/dialog"
+import { Dialog, DialogContent } from "@/ui/Dialog/dialog"
 import { Drawer, DrawerContent, DrawerOverlay } from "@/ui/drawer"
+import { cva } from "cva"
 import {
   ComponentProps,
   FC,
@@ -14,7 +14,7 @@ import {
 import { OneModalContent } from "./OneModalContent/OneModalContent"
 import { OneModalHeader } from "./OneModalHeader/OneModalHeader"
 import { OneModalProvider } from "./OneModalProvider"
-import { ContentPadding, ModalPosition } from "./types"
+import { ContentPadding, ModalPosition, ModalWidth } from "./types"
 import { useIsSmallScreen } from "./utils"
 
 export type OneModalProps = {
@@ -26,7 +26,11 @@ export type OneModalProps = {
   asBottomSheetInMobile?: boolean
   /** the padding of internal content areas (header, content, footer) */
   contentPadding?: ContentPadding
+
+  /** The position of the modal */
   position?: ModalPosition
+  /** The width of the modal. Only applies to center position but we can NOT use narrowing as position undefined is valid */
+  width?: ModalWidth
   /** Custom content to render in the modal. Only accepts OneModal.Header and OneModal.Content components */
   children:
     | ReactElement<
@@ -39,6 +43,60 @@ export type OneModalProps = {
       >[]
 } & Partial<Pick<TabsProps, "tabs" | "activeTabId" | "setActiveTabId">>
 
+const modalWrapperClassName = cva({
+  variants: {
+    variant: {
+      bottomSheet: "max-h-[95vh] bg-f1-background",
+      sidePosition:
+        "absolute bottom-3 top-3 flex w-full translate-x-0 translate-y-0 flex-col rounded-md border border-solid border-f1-border-secondary",
+      center: "flex",
+      fullscreen: "",
+    },
+    position: {
+      right: "left-auto right-3 items-end",
+      left: "left-3 items-start",
+      center: "",
+      fullscreen: "inset-6",
+    },
+  },
+  defaultVariants: {
+    variant: "center",
+  },
+})
+
+const modalContentClassName = cva({
+  variants: {
+    variant: {
+      bottomSheet: "max-h-[95vh] bg-f1-background",
+      sidePosition:
+        "flex h-full w-full flex-col rounded-md border border-solid border-f1-border-secondary",
+      center: "flex max-h-[95%] min-w-[100px] flex-1 flex-col rounded-xl",
+      fullscreen: "h-full w-full rounded-xl",
+    },
+    position: {
+      left: "",
+      right: "",
+      center: "",
+      fullscreen: "",
+    },
+    width: {
+      sm: "max-w-[539px]",
+      md: "max-w-[680px]",
+      lg: "max-w-[820px]",
+    },
+  },
+  compoundVariants: [
+    {
+      variant: "fullscreen",
+      width: ["sm", "md", "lg"],
+      class: "max-w-full",
+    },
+  ],
+  defaultVariants: {
+    variant: "center",
+  },
+})
+
 export const OneModal: FC<OneModalProps> = ({
   asBottomSheetInMobile = true,
   position = "center",
@@ -46,6 +104,7 @@ export const OneModal: FC<OneModalProps> = ({
   isOpen,
   contentPadding = "md",
   children,
+  width = "md",
 }) => {
   // Use state to store the container element so we can trigger re-renders
   // when it's set. This ensures child components like F0Select get the
@@ -72,25 +131,40 @@ export const OneModal: FC<OneModalProps> = ({
 
   const isSidePosition = position === "left" || position === "right"
 
-  const contentClassName = useMemo(() => {
-    if (isSmallScreen && asBottomSheetInMobile) {
-      return "max-h-[95vh] bg-f1-background"
+  const variant = useMemo(() => {
+    if (position === "fullscreen") {
+      return "fullscreen"
     }
-
+    if (isSmallScreen && asBottomSheetInMobile) {
+      return "bottomSheet"
+    }
     if (isSidePosition) {
-      return cn(
-        "w-full flex flex-col absolute top-3 bottom-3 translate-y-0 translate-x-0 max-w-[539px] rounded-md border border-solid border-f1-border-secondary",
-        position === "left" && "left-3",
-        position === "right" && "left-auto right-3"
+      return "sidePosition"
+    }
+    return "center"
+  }, [isSmallScreen, asBottomSheetInMobile, isSidePosition, position])
+
+  // Forces the width to be "sm" for sidePosition variants
+  const localWidth = useMemo(() => {
+    if (variant === "sidePosition") {
+      return "sm"
+    }
+    if (width && position !== "center") {
+      console.warn(
+        "OneModal: `width` prop is only applicable to center position"
       )
     }
 
-    if (position === "fullscreen") {
-      return "w-[calc(100%-48px)] h-[calc(100%-48px)] rounded-xl"
-    }
+    return width
+  }, [variant, width, position])
 
-    return "flex flex-col max-h-[620px] max-w-[680px] rounded-xl"
-  }, [position, isSmallScreen, asBottomSheetInMobile, isSidePosition])
+  const contentClassName = useMemo(() => {
+    return modalContentClassName({
+      variant,
+      position,
+      width: localWidth,
+    })
+  }, [variant, position, localWidth])
 
   if (isSmallScreen && asBottomSheetInMobile) {
     return (
@@ -129,7 +203,11 @@ export const OneModal: FC<OneModalProps> = ({
       >
         <DialogContent
           ref={setContentRef}
-          withTraslateAnimation={!isSidePosition}
+          withTranslateAnimation={!isSidePosition}
+          wrapperClassName={modalWrapperClassName({
+            variant,
+            position,
+          })}
           className={contentClassName}
         >
           {children}