+++ packages/react/src/ui/Dialog/components/DialogHeader.tsx
@@ -1,0 +1,7 @@
+import { HTMLAttributes } from "react"
+
+export const DialogHeader = ({
+  className,
+  ...props
+}: HTMLAttributes<HTMLDivElement>) => <div className={className} {...props} />
+DialogHeader.displayName = "DialogHeader"