+++ packages/react/src/ui/InputField/InputField.tsx
@@ -501,7 +501,7 @@ const InputField = forwardRef<HTMLDivElement, InputFieldProps<string>>(
             {!noEdit && (
               <div
                 className={cn(
-                  "pointer-events-none absolute bottom-0 left-0 top-[1px] z-10 flex flex-1 justify-start px-3 text-f1-foreground-secondary transition-opacity",
+                  "pointer-events-none absolute bottom-0 left-0 top-[1px] z-10 flex flex-1 justify-start px-3 text-f1-foreground-tertiary transition-opacity",
                   (icon || avatar) && "pl-8",
                   (icon || avatar) && size === "md" && "pl-9",
                   inputElementVariants({ size }),
@@ -509,7 +509,7 @@ const InputField = forwardRef<HTMLDivElement, InputFieldProps<string>>(
                     !hidePlaceholder &&
                     isEmpty(localValue) &&
                     !isAutofilled
-                    ? "opacity-1"
+                    ? "opacity-100"
                     : "opacity-0"
                 )}
                 onClick={handleClickPlaceholder}